Get a basic React version working
