+++ manifest.json
@@ -4,7 +4,7 @@
     "version": "1.0",
     "manifest_version": 2,
     "chrome_url_overrides" : {
-      "newtab": "index.html"
+      "newtab": "react-version/build/index.html"
     },
     "browser_action": {
       "default_icon": "icons/48.png"
@@ -15,3 +15,4 @@
       "128": "icons/128.png"
     }
  }
+
