+++ react-version/public/manifest.json
@@ -1,15 +1,12 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
-  "icons": [
-    {
-      "src": "favicon.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/x-icon"
+    "name": "I am Groot",
+    "description" : "It's a super fancy new tab extension",
+    "version": "1.0",
+    "manifest_version": 2,
+    "chrome_url_overrides" : {
+      "newtab": "index.html"
+    },
+    "browser_action": {
+      "default_icon": "favicon.ico"
     }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
+ }
