+++ react-version/src/config.json
