+++ react-version/src/index.css
@@ -1,14 +1,20 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+html,
+body,
+.BodyWrapper {
+  box-sizing: border-box;
+  height: 100%;
+  width: 100%;
+  font-size: 16px;
+  margin: 0px;
+  padding: 0px;
 }
 
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
+*, *:before, *:after {
+  box-sizing: inherit;
+}
+
+@font-face {
+  font-family: "Eksell Display Medium";
+  src: url(./assets/fonts/EksellDisplay-Small.woff) format('woff');
+  font-style: normal;
 }
